fix compiling warning for conversion

+++ src/exemodel/clientee.cpp
@@ -53,7 +53,7 @@ void clientee::dispose(poller & mgr, uint32_t evts)
 size_t clientee::recv(void * buffer, size_t length)
 {
 	ssize_t ret = ::recv(_fd(), buffer, length, 0);
-	validate_ret(ret, "socket recv");
+	validate_ret((int)ret, "socket recv");
 
 	return (size_t)ret;
 }
@@ -61,7 +61,7 @@ size_t clientee::recv(void * buffer, size_t length)
 size_t clientee::send(const void * buffer, size_t length)
 {
 	ssize_t ret = ::send(_fd(), buffer, length, 0);
-	validate_ret(ret, "socket send");
+	validate_ret((int)ret, "socket send");
 
 	return (size_t)ret;
 }
