+++ src/exemodel/connectee.cpp
@@ -27,7 +27,7 @@ void connectee::dispose(poller & mgr, uint32_t evts)
 size_t connectee::recv(void * buffer, size_t length)
 {
 	ssize_t ret = ::recv(_fd(), buffer, length, 0);
-	validate_ret(ret, "socket recv");
+	validate_ret((int)ret, "socket recv");
 
 	return (size_t)ret;
 }
@@ -35,7 +35,7 @@ size_t connectee::recv(void * buffer, size_t length)
 size_t connectee::send(const void * buffer, size_t length)
 {
 	ssize_t ret = ::send(_fd(), buffer, length, 0);
-	validate_ret(ret, "socket send");
+	validate_ret((int)ret, "socket send");
 
 	return (size_t)ret;
 }
